[VQueues] Create vqueue on state mutation when missing
ExternalStateMutation (admin API POST /services/{name}/state) was silently
dropped when the target virtual-object key had never been invoked, because
vqueues were only created lazily on the invocation path. Route state mutations
through a new VQueue::vqueue_for_state_mutation helper that get-or-creates the
vqueue with the same Lock link an exclusive invocation would produce.
