+++ crates/worker/src/partition/state_machine/mod.rs
@@ -113,7 +113,7 @@ use restate_types::vqueues::{self, EntryId, VQueueId};
 use restate_types::{RESTATE_VERSION_1_6_0, journal_v2};
 use restate_types::{RestateVersion, SemanticRestateVersion};
 use restate_types::{Versioned, journal::*};
-use restate_vqueues::{VQueue, VQueuesMetaCache, generate_vqueue_id};
+use restate_vqueues::{VQueue, VQueuesMetaCache};
 use restate_wal_protocol::Command;
 use restate_wal_protocol::timer::TimerKeyDisplay;
 use restate_wal_protocol::timer::TimerKeyValue;
@@ -5296,32 +5296,16 @@ impl<S> StateMachineApplyContext<'_, S> {
         let now = UniqueTimestamp::from_unix_millis_unchecked(self.record_created_at);
         let service_id = &state_mutation.service_id;
 
-        let qid = generate_vqueue_id(
-            service_id.partition_key(),
-            // We don't pass the scope here yet
-            &None,
-            // we don't pass the limit key here yet
-            &LimitKey::None,
-            true, /* is_exclusive */
-            &service_id.service_name,
-            Some(&service_id.key),
-        );
-
-        // todo: Make this a use-facing ID, generated at ingress.
+        // todo: Make this a user-facing ID, generated at ingress.
         let mutation_id = StateMutationId::generate(service_id.partition_key());
-        let Some(mut vqueue) = VQueue::get(
-            &qid,
+        let (qid, mut vqueue) = VQueue::vqueue_for_state_mutation(
+            now,
+            service_id,
             self.storage,
             self.vqueues_cache,
             self.is_leader.then_some(self.action_collector),
         )
-        .await?
-        else {
-            error!("State mutation request was ignored because the vqueue {qid} does not exist!");
-            // todo: When/if we made state mutations rpc-like, we should return the error to the
-            // user here.
-            return Ok(());
-        };
+        .await?;
 
         let entry_id = EntryId::from(mutation_id);
         vqueue.enqueue_new(
